# Conflicts:
#	src/Typewriter/Typewriter.csproj
#	src/Typewriter/VisualStudio/ExtensionPackage.cs
#	src/Typewriter/packages.config
#	src/Typewriter/source.extension.vsixmanifest

+++ src/CodeModel/CodeModel/EnumValue.cs
@@ -41,7 +41,7 @@ namespace Typewriter.CodeModel
         /// <summary>
         /// The numeric value.
         /// </summary>
-        public abstract int Value { get; }
+        public abstract long Value { get; }
 
         /// <summary>
         /// Converts the current instance to string.
